chore: split up the app state file for clarity
This file was overloaded with implementations.
Now that there are reducers that operate over
limited state and action, this pulls them out into
their own dedicated file.

Better clarity and easier to follow too!
