+++ Watch Extension/State Management/PulseReducer.swift
@@ -1,0 +1,12 @@
+import Swift
+
+enum PulseAction {
+    case timerPulse
+}
+
+func pulseReducer(timerPulseCount: inout Int, action: PulseAction) {
+    switch action {
+    case .timerPulse:
+        timerPulseCount += 1
+    }
+}
