+++ Watch Extension/State Management/SavedFyiDialogReducer.swift
@@ -1,0 +1,12 @@
+import Swift
+
+enum SavedFyiDialogAction {
+    case hideSavedFyiDialog
+}
+
+func savedFyiDialogReducer(showSavedFyiDialog: inout Bool, action: SavedFyiDialogAction) {
+    switch action {
+    case .hideSavedFyiDialog:
+        showSavedFyiDialog = false
+    }
+}
